third_party/blink/web_tests/fast/events/touch/touch-action-preventDefault-expected.txt

CONSOLE ERROR: Unable to preventDefault inside passive event listener due to target being treated as passive. See https://www.chromestatus.com/feature/5093566007214080
This is a testharness.js-based test.
Harness: the test ran to completion.
